+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-### [0.0.0] - 9/11/20
-
-- Site is now live at smartshuffle.io
+### [0.2.0] - 9/16/20
+- Added about page.
+- Added privacy policy.
+- Added changelog to the site.
+- Added contact us.
+- Reworked the main navbar, content is now centered.
 
 ### [0.0.1] - 9/12/20
 
@@ -10,3 +13,7 @@
 - Added an idle timer that will cause the site to stop making now playing API requests after 5 minutes of inactivity.
 - Local tracks are now displayed in now playing correctly.
 - Fixed a bug with edit playlist weights that caused an error to always appear, rendering the feature unusable.
+
+### [0.0.0] - 9/11/20
+
+- Site is now live at smartshuffle.io